requirements.txt
# ===== RAG System & Memory =====
chromadb==1.3.0
sentence-transformers==5.1.2
openai>=1.0.0

# ===== Core Utilities =====
python-dotenv==1.0.0
pydantic>=2.7.0
typer>=0.12.5
click==8.1.7
rich==13.7.0

# ===== Database =====
mysql-connector-python==8.1.0

# ===== Bank Statement Parsing =====
pandas>=2.1.4
docling>=2.0.0
ofxparse==0.21

# ===== Duplicate Detection =====
rapidfuzz

# ===== Data Validation & Processing =====
python-dateutil==2.8.2
charset-normalizer==3.3.2

# ===== CLI & User Interface =====
tabulate==0.9.0

# ===== Testing =====
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# ===== Development Tools =====
black==23.10.1
flake8==6.1.0
mypy==1.6.1
pre-commit==3.5.0

# ===== Logging & Monitoring =====
structlog==23.2.0
python-json-logger==2.0.7

# ===== Web Interface (Optional) =====
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# ===== Data Export =====
xlsxwriter==3.1.9
reportlab==4.0.7

# ===== Configuration =====
pyyaml==6.0.1
pgvector
